+++ examples/transformer/gpt.py
@@ -1,0 +1,85 @@
+import sys
+import os
+sys.path.insert(0, os.path.join(os.path.dirname(__file__), '../..'))
+
+import numpy as np
+import forgeNN as fnn
+from forgeNN.nn import set_seed
+
+# hyperparameters
+batch_size = 64 # how many independent sequences will we process in parallel?
+block_size = 256 # what is the maximum context length for predictions?
+max_iters = 5000
+eval_interval = 500
+learning_rate = 3e-4
+device = 'cuda' if fnn.is_cuda_available() else 'cpu'
+eval_iters = 200
+n_embd = 384
+n_head = 6
+n_layer = 6
+dropout = 0.2
+# ------------
+set_seed(1337)
+
+# 1. Load and process the dataset
+with open('examples/transformer/input.txt', 'r', encoding='utf-8') as f:
+    text = f.read()
+
+# here are all the unique characters that occur in this text
+chars = sorted(list(set(text)))
+vocab_size = len(chars)
+# create a mapping from characters to integers
+stoi = { ch:i for i,ch in enumerate(chars) }
+itos = { i:ch for i,ch in enumerate(chars) }
+encode = lambda s: [stoi[c] for c in s] # encoder: take a string, output a list of integers
+decode = lambda l: ''.join([itos[i] for i in l]) # decoder: take a list of integers, output a string
+
+
+print(encode("hii there"))
+print(decode(encode("hii there")))
+
+# Train and test splits
+tensor = fnn.Tensor(np.array(encode(text)))
+data = tensor.data
+print(tensor.shape)
+print(data[:1000])
+
+# Let's now split up the data into train and validation sets
+n = int(0.9*len(data)) # first 90% will be train, rest val
+train_data = data[:n]
+val_data = data[n:]
+
+train_data[:block_size+1]
+
+x = train_data[:block_size]
+y = train_data[1:block_size+1]
+for t in range(block_size):
+    context = x[:t+1]
+    target = y[t]
+    print(f"when input is {context} the target: {target}")
+
+
+# TODO: implement fnn.stack and fnn.randint to replace the following code
+# def get_batch(split):
+#     # generate a small batch of data of inputs x and targets y
+#     data = train_data if split == 'train' else val_data
+#     ix = torch.randint(len(data) - block_size, (batch_size,))
+#     x = torch.stack([data[i:i+block_size] for i in ix])
+#     y = torch.stack([data[i+1:i+block_size+1] for i in ix])
+#     return x, y
+
+# xb, yb = get_batch('train')
+# print('inputs:')
+# print(xb.shape)
+# print(xb)
+# print('targets:')
+# print(yb.shape)
+# print(yb)
+
+# print('----')
+
+# for b in range(batch_size): # batch dimension
+#     for t in range(block_size): # time dimension
+#         context = xb[b, :t+1]
+#         target = yb[b,t]
+#         print(f"when input is {context.tolist()} the target: {target}")